+++ ripozo/viewsets/relationships/__init__.py
@@ -3,4 +3,5 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-__author__ = 'Tim Martin'
+from ripozo.viewsets.relationships.list_relationship import ListRelationship
+from ripozo.viewsets.relationships.relationship import Relationship